requirements.txt
fastapi==0.109.1
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pymysql==1.1.0
alembic==1.12.0
httpx==0.25.0
pytest==7.4.2
email-validator==2.1.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1
Jinja2==3.1.3
PyJWT==2.8.0
python-multipart==0.0.7
